fix(guide-mode): missing spawn events
In guide mode, when a bot is added to a space without any guides it does not send a spawn event or call hears().  When a guide is added to the space the bot becomes active generating a spawn.  This wasn't happening if the membership email wasn't all lower case.